fix: 格式化

+++ .eslintrc.js
@@ -25,7 +25,7 @@ module.exports = {
     '@typescript-eslint/no-unused-vars': ['error', { vars: 'all', args: 'after-used' }],
     'vue/require-default-prop': 'off', // prop 默认值
     'vue/require-prop-types': 'off', // prop type
-    'vue/html-quotes': ['warn', 'single'],
+    'vue/html-quotes': ['warn', 'double'],
     'vue/max-attributes-per-line': 'off',
     'vue/singleline-html-element-content-newline': 'off',
     'vue/attributes-order': 'off',
